well, i was wrong

+++ code/modules/mob/inventory.dm
@@ -267,15 +267,19 @@
 	return 0
 
 //Helper proc tied to above for creating something in-hand via construction
-/mob/proc/create_in_hands(var/obj/item/olditem, var/obj/item/newitem, var/obj/item/using, var/must_hold = TRUE)
+/mob/proc/create_in_hands(var/obj/item/olditem, var/obj/item/newitem, var/obj/item/using, var/uses = 1)
 	if(!olditem || !newitem)
 		return 0
-	if(olditem.loc == src || !must_hold)
+	if(olditem.loc == src)
 		drop_item(olditem, force_drop = 1) // Necessary to show up in the same hand for below
 		. = put_in_hands(newitem)
 	qdel(olditem)
 	if(using)
-		qdel(using)
+		if(istype(using,/obj/item/stack) && uses)
+			var/obj/item/stack/S = using
+			S.use(uses)
+		else
+			qdel(using)
 
 /mob/proc/set_hand_amount(new_amount)
 	if(new_amount < held_items.len) //Decrease hand amount - drop items held in hands which will no longer exist!
